+++ R/api_azure_openai.R
@@ -32,7 +32,7 @@
 #' \dontrun{
 #' # Basic usage
 #' msg <- llm_message("What is R programming?")
-#' result <- azure_openai(msg)
+#' result <- azure_openai_chat(msg)
 #' 
 #' # With custom parameters
 #' result2 <- azure_openai_chat(msg,